# Issue 338: Use mutex for SMA locks