--- a/package.xml
+++ b/package.xml
@@ -126,6 +126,8 @@ and unpersisting values.
    <file name="apc_lock.h" role="src" />
    <file name="apc_mmap.c" role="src" />
    <file name="apc_mmap.h" role="src" />
+   <file name="apc_mutex.c" role="src" />
+   <file name="apc_mutex.h" role="src" />
    <file name="apc.php" role="src" />
    <file name="apc_php.h" role="src" />
    <file name="apc_pool_api.h" role="src" />