This has three parts:

- a) Add a mutex implementation
- b) Abstract out the lock implementation in SMA
- c) Use the mutex implementation in SMA

As written already in #338 benchmarks are favorable (20% faster):

Before: 8.795 seconds
After: 7.336 seconds

With less concurrency and requests (`ab -n40 -c5`):

Before: 4.8 seconds
After: 3.7 seconds
=> 29% faster

Fixes: #338