@LionsAd 2018-11-18
Thanks for the review, I'll fix those issues today.