@LionsAd 2018-11-19
I added a commit to disable the destroying of the lock for now.