@nikic 2018-11-19
Rebased and merged. Thanks a lot!